Smartfreelance_AlexIssa/projects/Smartfreelance_AlexIssa-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
requests==2.32.5
pydantic==2.10.3
pillow==10.4.0